refactor(code-factory): move test helper parsing into plan
